Deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
numpy
plotly
Pillow
scikit-learn==1.2.2